requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
lob>=4.5.4
python-dotenv>=1.0.0
